added time and cycle counting, estimate CPU clock from result and print it

+++ 6502.c
@@ -49,7 +49,7 @@ typedef struct microprocessor {
     unsigned char sp;
     unsigned short pc;
 	unsigned char status;
-    unsigned int cycles;
+    unsigned long cycles;
 } microprocessor;
 
 microprocessor cpu;
